feat(usda): improve parser coverage and add validation CLI
Parser fixes:
- Fix timecode fields to use Double instead of Uint64 (negative/fractional values)
- Add triple single-quoted string tokenization ('''...''')
- Support internal prim references (</Prim> without asset path)
- Handle None token for array values (returns empty arrays)
- Add generic fallback for unknown prim metadata

New features:
- Add validate_usd CLI binary for testing parser coverage
- Add cli feature flag with clap, rayon, indicatif, walkdir dependencies

Validation against usd-wg-assets improved from 76% to 92% pass rate.

+++ src/usda/parser/hierarchy.rs
@@ -49,13 +49,13 @@ impl<'a> super::Parser<'a> {
 
         const KNOWN_PROPS: &[(&str, Type)] = &[
             (FieldKey::DefaultPrim.as_str(), Type::Token),
-            (FieldKey::StartTimeCode.as_str(), Type::Uint64),
+            (FieldKey::StartTimeCode.as_str(), Type::Double),
             (FieldKey::HasOwnedSubLayers.as_str(), Type::StringVec),
             ("doc", Type::String),
-            ("endTimeCode", Type::Uint64),
-            ("framesPerSecond", Type::Uint64),
+            ("endTimeCode", Type::Double),
+            ("framesPerSecond", Type::Double),
             ("metersPerUnit", Type::Double),
-            ("timeCodesPerSecond", Type::Uint64),
+            ("timeCodesPerSecond", Type::Double),
             ("upAxis", Type::Token),
         ];
